requirements.txt
anndata==0.11.4
anyio==4.9.0
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
array-api-compat==1.12.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
biofluff==3.0.4
biopython==1.85
biothings_client==0.4.1
bleach==6.2.0
cached-property==1.5.2
celloracle==0.21.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.3
configparser==7.2.0
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.2
debugpy==1.8.15
decorator==5.2.1
defusedxml==0.7.1
diskcache==5.6.3
docopt==0.6.2
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.0
fastjsonschema==2.21.1
feather-format==0.4.1
filelock==3.18.0
fonttools==4.59.0
fqdn==1.5.1
ftpretty==0.4.0
genomepy==0.16.2
gimmemotifs==0.17.2
goatools==1.4.12
h11==0.16.0
h5py==3.14.0
HTSeq==2.0.9
httpcore==1.0.9
httpx==0.28.1
idna==3.10
igraph==0.11.9
iniconfig==2.3.0
ipykernel==6.30.0
ipython==8.37.0
ipywidgets==8.1.7
isoduration==20.11.0
iteround==1.0.4
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter_client==8.6.3
jupyter-console==6.6.3
jupyter_core==5.8.1
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
lark==1.2.2
legacy-api-wrap==1.4.1
llvmlite==0.44.0
logomaker==0.8.7
loguru==0.7.3
loompy==3.0.8
louvain==0.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.7
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mygene==3.2.2
mysql-connector-python==9.4.0
natsort==8.4.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
norns==0.1.6
nose==1.3.7
notebook==7.4.4
notebook_shim==0.2.4
numba==0.61.2
numpy==1.26.4
numpy-groupies==0.11.3
openpyxl==3.1.5
overrides==7.7.0
packaging==25.0
palettable==3.3.3
pandas==1.5.3
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.3.0
pip==25.1.1
platformdirs==4.3.8
pluggy==1.6.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pybedtools==0.12.0
pyBigWig==0.3.24
pycparser==2.22
pydot==4.0.1
pyfaidx==0.8.1.4
Pygments==2.19.2
pynndescent==0.5.13
pyparsing==3.2.3
pysam==0.23.3
pytest==8.4.2
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
qnorm==0.9.0
referencing==0.36.2
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rpds-py==0.26.0
scanpy==1.11.4
scikit-learn==1.5.2
scipy==1.15.2
seaborn==0.13.2
Send2Trash==1.8.3
session-info2==0.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
statsmodels==0.14.5
terminado==0.18.1
texttable==1.7.0
threadpoolctl==3.6.0
tinycss2==1.4.0
tomli==2.2.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20250708
typing_extensions==4.14.1
tzdata==2025.2
umap-learn==0.5.7
uri-template==1.3.0
urllib3==2.5.0
velocyto==0.17.17
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.45.1
widgetsnbextension==4.0.14
xdg==6.0.0
xlsxwriter==3.2.5
xxhash==3.5.0
